requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.8
Django==3.2.7
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
Markdown==3.3.4
pycparser==2.20
pytz==2021.1
sqlparse==0.4.2
